Fix not passing enough memory to cd-hit
cd-hit requires at least 2000 mb of memory to be passed to it through the -M flag. The Cdhit.pm script reserves this amount for the job runner, but sets -M to 0.9*2000=1800 mb, which causes cd-hit to fail. This commit fixes the issue by allocating more memory in total (2225 mb => 0.9*2225=2002 mb).